refactor(editor): keep richTextVariable registration out of engine/registry.ts
Two cohesion improvements to keep the generic engine layer free of
rich-text-specific knowledge:

- Extract richTextVariable's component definition into
  `components/richTextVariable/richtextvar-registration.ts` exporting
  `createRichTextVariableDefinition()`, mirroring the pattern used by
  table / qrcode / columns / datatable / placeholder. `engine/registry.ts`
  now calls `registry.register(createRichTextVariableDefinition())` and
  drops two domain-specific imports (`richTextVariablePathDisabled`,
  side-effect import of `EpistolaRichTextVariablePreview`) and the
  `html` import that's no longer needed.

- `data-contract/sections/SchemaSection.ts` derives the rich-text ids
  in its `FIELD_TYPES` and `ARRAY_ITEM_TYPES` arrays from
  `REF_TYPES.map(t => t.id)` instead of listing `'richTextInline'` /
  `'richTextBlock'` literally. Adding a new ref type now extends both
  dropdowns automatically.

No behaviour change. All 1500 editor tests still pass; the component
registry still emits 19 descriptors with `richTextVariable` present.
